@@ -1,6 +1,6 @@
 {
   "name": "dtool-lookup-webapp",
-  "version": "0.3.0",
+  "version": "0.4.1",
   "private": true,
   "scripts": {
     "serve": "vue-cli-service serve",
@@ -8,25 +8,29 @@
     "lint": "vue-cli-service lint"
   },
   "dependencies": {
-    "axios": "^0.18.0",
-    "bootstrap": "^4.3.1",
-    "bootstrap-vue": "^2.0.0-rc.22",
-    "core-js": "^2.6.5",
-    "filesize": "^4.1.2",
+    "@soerenmartius/vue3-clipboard": "^0.1.2",
+    "@vue/compat": "^3.2.31",
+    "axios": "^1.1.3",
+    "bootstrap": "^5.1.3",
+    "bootstrap-vue-3": "^0.3.12",
+    "core-js": "^3.21.1",
+    "filesize": "^8.0.7",
     "json2yaml": "^1.1.0",
     "moment": "^2.24.0",
-    "vue": "^2.6.10",
-    "vue-clipboard2": "^0.3.0",
-    "vuex": "^3.0.1"
+    "vue": "^3.2.31",
+    "vue-axios": "^3.4.1",
+    "vuex": "^4.0.2"
   },
   "devDependencies": {
-    "@vue/cli-plugin-babel": "^3.7.0",
-    "@vue/cli-plugin-eslint": "^3.7.0",
-    "@vue/cli-service": "^4.2.2",
-    "@vue/eslint-config-prettier": "^4.0.1",
-    "babel-eslint": "^10.0.1",
-    "eslint": "^5.16.0",
-    "eslint-plugin-vue": "^5.0.0",
-    "vue-template-compiler": "^2.5.21"
+    "@babel/eslint-parser": "^7.17.0",
+    "@babel/preset-react": "^7.16.7",
+    "@vue/cli-plugin-babel": "^5.0.3",
+    "@vue/cli-plugin-eslint": "^5.0.4",
+    "@vue/cli-service": "^5.0.4",
+    "@vue/compiler-sfc": "^3.2.31",
+    "@vue/eslint-config-prettier": "^7.0.0",
+    "eslint": "^8.11.0",
+    "eslint-plugin-vue": "^9.6.0",
+    "webpack": "^5.7.0"
   }
 }
