@@ -1,19 +1,13 @@
 <template>
   <div>
     <h3>Readme</h3>
-    <pre>{{ yaml.stringify(readme) }}</pre>
+    <pre>{{ readme }}</pre>
   </div>
 </template>
 
 <script>
-var yaml = require("json2yaml");
 export default {
-  name: "Readme",
-  data: function() {
-    return {
-      yaml: yaml
-    };
-  },
+  name: "DatasetReadme",
   computed: {
     readme: function() {
       return this.$store.state.current_dataset_readme;
